+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/Helpers/FileTransferable.java
@@ -1,4 +1,4 @@
-package org.vibehistorian.midimasterpiece.midigenerator;
+package org.vibehistorian.midimasterpiece.midigenerator.Helpers;
 
 import java.awt.datatransfer.DataFlavor;
 import java.awt.datatransfer.Transferable;
