+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/Panels/DrumGenSettings.java
@@ -1,4 +1,4 @@
-package org.vibehistorian.midimasterpiece.midigenerator;
+package org.vibehistorian.midimasterpiece.midigenerator.Panels;
 
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
