+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/Parts/ArpPart.java
@@ -1,4 +1,4 @@
-package org.vibehistorian.midimasterpiece.midigenerator;
+package org.vibehistorian.midimasterpiece.midigenerator.Parts;
 
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
